more progress on the delegating connection
since multiple things will need to access the decorated internal
connection at the same time (potentially), we'll need a mutex and an arc
and locking here. womp womp I was trying to avoid that, but I don't
think it's possible to have concurrent mutable access without
synchronization, and I think I need concurrent mutable access
