+++ src/caseta/mod.rs
@@ -1,4 +1,3 @@
 pub mod connection;
-pub mod connectionmanager;
 pub mod message;
 pub mod remote;
