+++ src/main.rs
@@ -3,6 +3,10 @@ use std::collections::HashMap;
 use std::sync::Arc;
 
 use anyhow::{anyhow, Result};
+use caseta_listener::caseta::connection::{
+    CasetaConnectionProvider, DefaultCasetaConnectionProvider, DefaultTcpSocketProvider,
+    DelegatingCasetaConnectionManager, ReadOnlyConnection,
+};
 use caseta_listener::client::room_state::new_cache;
 use tokio::sync::mpsc;
 use tracing::subscriber::set_global_default;
@@ -11,9 +15,6 @@ use tracing_bunyan_formatter::{BunyanFormattingLayer, JsonStorageLayer};
 use tracing_subscriber::layer::SubscriberExt;
 use tracing_subscriber::{EnvFilter, Registry};
 
-use caseta_listener::caseta::connection::{
-    CasetaConnectionError, CasetaConnectionManager, DefaultTcpSocketProvider,
-};
 use caseta_listener::caseta::message::Message;
 use caseta_listener::caseta::remote::{remote_watcher_loop, RemoteWatcher};
 use caseta_listener::client::dispatcher::{dispatcher_loop, DeviceActionDispatcher};
@@ -52,13 +53,16 @@ async fn watch_caseta_events() -> Result<()> {
 
     let caseta_address = auth_configuration.caseta_host.clone();
     let port = auth_configuration.caseta_port;
-    let tcp_socket_provider = DefaultTcpSocketProvider::new(caseta_address, port);
-    let mut connection = CasetaConnectionManager::new(
+
+    let tcp_socket_provider = Box::new(DefaultTcpSocketProvider::new(caseta_address, port));
+    let connection_manager_provider = DefaultCasetaConnectionProvider::new(
+        caseta_address,
+        port,
         auth_configuration.caseta_username,
         auth_configuration.caseta_password,
-        &tcp_socket_provider,
+        tcp_socket_provider,
     );
-    connection.initialize().await?;
+    let connection = DelegatingCasetaConnectionManager::new(Box::new(connection_manager_provider));
 
     let (action_sender, action_receiver) = mpsc::channel(64);
     let mut remote_watchers: RemoteWatcherDb = HashMap::new();
